Backend/requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.3.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
requests==2.27.1
python-dotenv==0.20.0
django-meta==2.0.0
django-imagekit==4.0.2
django-sitemap==0.6.0
